# src/main/java/toby/ai/tobyreminder/dto/request/ReminderRequest.java
package toby.ai.tobyreminder.dto.request;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class ReminderRequest {

    private String title;
    private String notes;
    private Long listId;
}